// server/enums/dpsRole.ts
enum DpsRole {
  // SAN roles
  ROLE_SAN_EDITOR = 'ROLE_SAN_EDITOR',
  ROLE_SAN_EDUCATION_MANAGER = 'ROLE_SAN_EDUCATION_MANAGER',

  // DPR Reporting
  ROLE_PRISONS_REPORTING_USER = 'ROLE_PRISONS_REPORTING_USER',

  // Other DPS roles
  ROLE_GLOBAL_SEARCH = 'ROLE_GLOBAL_SEARCH',
  ROLE_INACTIVE_BOOKINGS = 'ROLE_INACTIVE_BOOKINGS',
  ROLE_POM = 'ROLE_POM',
}

export default DpsRole