// server/enums/applicationAction.ts
enum ApplicationAction {
  SEARCH = 'SEARCH',
  VIEW_PROFILE = 'VIEW_PROFILE',
  RECORD_EDUCATION_LEARNER_SUPPORT_PLAN = 'RECORD_EDUCATION_LEARNER_SUPPORT_PLAN',
  UPDATE_EDUCATION_LEARNER_SUPPORT_PLAN = 'UPDATE_EDUCATION_LEARNER_SUPPORT_PLAN',
  RECORD_CHALLENGES = 'RECORD_CHALLENGES',
  RECORD_STRENGTHS = 'RECORD_STRENGTHS',
  RECORD_SELF_DECLARED_CONDITIONS = 'RECORD_SELF_DECLARED_CONDITIONS',
  RECORD_DIAGNOSED_CONDITIONS = 'RECORD_DIAGNOSED_CONDITIONS',
}

export default ApplicationAction